Remove pytest-lazy-fixture
