+++ images/plbase/python-requirements.txt
@@ -26,7 +26,6 @@ pygments-ansi-color==0.3.0
 pygraphviz==1.6
 pyquaternion==0.9.9
 pytest-cov==4.1.0
-pytest-lazy-fixture==0.6.3
 pytest-repeat==0.9.3
 pytest==7.4.4
 recommonmark==0.7.1
@@ -43,6 +42,5 @@ text-unidecode==1.3
 types-chevron==0.14.2.20240106
 types-lxml==2023.10.21
 types-Pygments==2.17.0.20240106
-types-pytest-lazy-fixture==0.6.3.4
 typing-extensions==4.9.0
 tzlocal==4.3
